refactor: moved weatherforecast default to correct dir
