+++ moutarde-back/moutarde-back/Entities/WeatherForecast.cs
@@ -1,4 +1,4 @@
-namespace moutarde_back;
+namespace moutarde_back.Entities;
 
 public class WeatherForecast
 {
