+++ moutarde-back/moutarde-back/Features/Temporary/WeatherForecastController.cs
@@ -1,4 +1,5 @@
 using Microsoft.AspNetCore.Mvc;
+using moutarde_back.Entities;
 
 namespace moutarde_back.Features.Temporary;
